refactor: migrate helper classes to extension functions with documentation
