--- pushnotification-local/src/main/java/io/github/mikan/pushnotification/local/NotificationHelper.kt
@@ -1,63 +1,0 @@
-package io.github.mikan.pushnotification.local
-
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import androidx.core.app.NotificationCompat
-
-object NotificationHelper {
-    const val CHANNEL_ID = "local_notification_channel"
-    private const val CHANNEL_NAME = "ローカル通知"
-    private const val CHANNEL_DESCRIPTION = "スケジュールされたローカル通知"
-
-    fun createNotificationChannel(context: Context) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                CHANNEL_ID,
-                CHANNEL_NAME, NotificationManager.IMPORTANCE_HIGH
-            ).apply {
-                description = CHANNEL_DESCRIPTION
-                enableVibration(true)
-                enableLights(true)
-            }
-
-            context.notificationManager.createNotificationChannel(channel)
-        }
-    }
-
-    fun showNotification(
-        context: Context,
-        notificationId: Int,
-        title: String,
-        content: String,
-        targetActivity: Class<*>? = null
-    ) {
-        val intent = targetActivity?.let {
-            Intent(context, it).apply {
-                flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-            }
-        }
-
-        val pendingIntent = intent?.let {
-            PendingIntent.getActivity(
-                context,
-                0,
-                it,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-            )
-        }
-
-        context.notify(CHANNEL_ID, notificationId) {
-            setSmallIcon(R.drawable.ic_notification)
-            setContentTitle(title)
-            setContentText(content)
-            setStyle(NotificationCompat.BigTextStyle().bigText(content))
-            setPriority(NotificationCompat.PRIORITY_HIGH)
-            pendingIntent?.let { setContentIntent(it) }
-            setAutoCancel(true)
-        }
-    }
-}
